+++ src/Vizy.php
@@ -72,7 +72,7 @@ class Vizy extends Plugin
     // Protected Methods
     // =========================================================================
 
-    protected function createSettingsModel(): ?Model
+    protected function createSettingsModel(): Settings
     {
         return new Settings();
     }