+++ src/base/Routes.php
@@ -14,6 +14,7 @@ trait Routes
     public function _registerCpRoutes(): void
     {
         Event::on(UrlManager::class, UrlManager::EVENT_REGISTER_CP_URL_RULES, function(RegisterUrlRulesEvent $event): void {
+            $event->rules['vizy'] = 'vizy/settings/index';
             $event->rules['vizy/settings'] = 'vizy/settings/index';
         });
     }