+++ src/elements/Block.php
@@ -1,7 +1,9 @@
 <?php
 namespace verbb\vizy\elements;
 
+use verbb\vizy\fields\VizyField;
 use verbb\vizy\helpers\Matrix;
+use verbb\vizy\models\BlockType;
 
 use craft\base\Element;
 use craft\models\FieldLayout;
@@ -16,12 +18,19 @@ class Block extends Element
         return true;
     }
 
+    public static function isLocalized(): bool
+    {
+        return true;
+    }
+
 
     // Properties
     // =========================================================================
 
     private ?FieldLayout $_fieldLayout = null;
     private mixed $_owner = null;
+    private ?BlockType $_type = null;
+    private ?VizyField $_field = null;
 
 
     // Public Methods
@@ -32,9 +41,29 @@ class Block extends Element
         return $this->_fieldLayout;
     }
 
-    public function setFieldLayout($fieldLayout): void
+    public function setFieldLayout(?FieldLayout $fieldLayout): void
     {
         $this->_fieldLayout = $fieldLayout;
+    }
+
+    public function getType(): ?BlockType
+    {
+        return $this->_type;
+    }
+
+    public function setType(?BlockType $type): void
+    {
+        $this->_type = $type;
+    }
+
+    public function getField(): ?VizyField
+    {
+        return $this->_field;
+    }
+
+    public function setField(?VizyField $field): void
+    {
+        $this->_field = $field;
     }
 
     public function setFieldValues(array $values): void