+++ src/web/assets/field/src/js/components/VizyInput.vue
@@ -1,6 +1,6 @@
 <template>
     <div>
-        <div v-if="editor" class="vui-rich-text" :class="{ 'has-focus': isFocused() }">
+        <div v-if="editor" class="vui-rich-text" :class="{ 'has-focus': isFocused() }" :style="{ '--rows': settings.initialRows }">
             <menu-bar v-if="buttons.length" ref="toolbar" :buttons="buttons" :editor="editor" :field="this" />
             <code-editor v-model="codeEditorHtml" :visible="showCodeEditor" :editor="editor" :field="this" />
             <editor-content :class="{ 'code-view': showCodeEditor }" class="vui-editor" :editor="editor" />
@@ -171,11 +171,11 @@ export default {
                 window.addEventListener('resize', this.updateFixedToolbar);
 
                 // Handle the element editor slideout
-                const slideout = document.querySelector('.slideout[data-element-editor] .so-body');
+                const $slideout = document.querySelector('.slideout[data-element-editor].so-visible .so-body');
 
-                if (slideout) {
-                    slideout.addEventListener('scroll', this.updateFixedToolbarEditor);
-                    slideout.addEventListener('resize', this.updateFixedToolbarEditor);
+                if ($slideout) {
+                    $slideout.addEventListener('scroll', this.updateFixedToolbarEditor);
+                    $slideout.addEventListener('resize', this.updateFixedToolbarEditor);
                 }
 
                 Garnish.on(Craft.Preview, 'open', this.openLivePreviewCallback);
@@ -397,6 +397,14 @@ export default {
 
         openLivePreviewCallback() {
             this.isLivePreview = true;
+
+            // Handle the Live Preview scroll
+            const $livePreview = document.querySelector('.lp-editor-container .lp-editor');
+
+            if ($livePreview) {
+                $livePreview.addEventListener('scroll', this.updateFixedToolbar);
+                $livePreview.addEventListener('resize', this.updateFixedToolbar);
+            }
         },
 
         closeLivePreviewCallback() {
@@ -407,7 +415,7 @@ export default {
             let headerBuffer = document.querySelector('body.fixed-header #header') ? document.querySelector('body.fixed-header #header').offsetHeight : 0;
 
             if (this.isLivePreview) {
-                headerBuffer = document.querySelector('.lp-editor-container header.flex') ? document.querySelector('.lp-editor-container header.flex').offsetHeight : 0;
+                headerBuffer = document.querySelector('.lp-editor-container header.flex') ? document.querySelector('.lp-editor-container header.flex').offsetHeight - parseFloat(window.getComputedStyle(document.querySelector('.lp-editor-container .lp-editor'), null).getPropertyValue('padding-top')) : 0;
             }
 
             // Apply any parent Vizy fields toolbars, otherwise we get multiple toolbar overlaps
@@ -482,6 +490,8 @@ export default {
 // ==========================================================================
 
 .vui-rich-text {
+    --rows: 7;
+
     position: relative;
     border-radius: 3px;
     border: 1px solid rgba(96, 125, 159, 0.25);
@@ -496,7 +506,7 @@ export default {
         outline: none;
         word-wrap: normal;
         padding: 16px;
-        min-height: 10rem;
+        min-height: calc(2rem + (var(--rows) * 1rem));
         background-color: #fbfcfe;
         background-clip: padding-box;
 
@@ -563,6 +573,12 @@ export default {
         text-decoration: underline;
     }
 
+    .h1, .h2, .h3, .h4, h1, h2, h3, h4 {
+        font-weight: 700;
+        line-height: 1.2;
+        margin: 14px 0;
+    }
+
     &.code-view {
         opacity: 0;
         visibility: hidden;
@@ -602,6 +618,11 @@ export default {
 // Fix Redactor incompatibility with ProseMirror style
 .ProseMirror [contenteditable="false"] .redactor [contenteditable="true"] {
     white-space: normal;
+}
+
+// Fix Live Preview overlapping
+.lp-editor-container > header {
+    z-index: 100;
 }
 
 </style>